Changed double to float

+++ WebApiCT/Entities/DataTransferObjects/ExerciseForCreateDto.cs
@@ -9,6 +9,6 @@ namespace CaloriesTracker.Entities.DataTransferObjects
         [Required(ErrorMessage = "Description is required field.")]
         public string Description { get; set; }
         [Range(0.01, 10.0, ErrorMessage = "CaloriesSpent must be between 0.01 and 10.0")]
-        public double CaloriesSpent { get; set; }
+        public float CaloriesSpent { get; set; }
     }
 }
