+++ WebApiCT/Entities/DataTransferObjects/IngredientEatingForCreateDto.cs
@@ -8,6 +8,6 @@ namespace CaloriesTracker.Entities.DataTransferObjects
         [Required(ErrorMessage = "IngredientId is required field.")]
         public Guid IngredientId { get; set; }
         [Range(0.01, 10000, ErrorMessage = "Grams bust be betwewn 0.01 and 10000.0")]
-        public double Grams { get; set; }
+        public float Grams { get; set; }
     }
 }
