+++ WebApiCT/Entities/DataTransferObjects/IngredientRecipeForReadDto.cs
@@ -5,6 +5,6 @@ namespace CaloriesTracker.Entities.DataTransferObjects
     public class IngredientRecipeForReadDto
     {
         public Guid IngredientId { get; set; }
-        public double Grams { get; set; }
+        public float Grams { get; set; }
     }
 }
