+++ WebApiCT/Entities/Models/Exercise.cs
@@ -8,7 +8,7 @@ namespace CaloriesTracker.Entities.Models
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double CaloriesSpent { get; set; }
+        public float CaloriesSpent { get; set; }
         public virtual IEnumerable<ActivityExercise> ActivityExercise { get; set; }
 
         public Exercise()
