+++ WebApiCT/Entities/Models/Ingredient.cs
@@ -7,10 +7,10 @@ namespace CaloriesTracker.Entities.Models
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public double Calories { get; set; }
-        public double Proteins { get; set; }
-        public double Fats { get; set; }
-        public double Carbohydrates { get; set; }
+        public float Calories { get; set; }
+        public float Proteins { get; set; }
+        public float Fats { get; set; }
+        public float Carbohydrates { get; set; }
         public virtual IEnumerable<IngredientEating> IngredientEating { get; set; }
         public virtual IEnumerable<IngredientRecipe> IngredientRecipe { get; set; }
